refactor: consolidate all resolution logic into live_docs_fetch.py
Move yaml_url, event_url and their helpers into live_docs_fetch.py,
eliminating live_docs.py and live_docs_queries_data.py entirely.
The server shims (BadRequest, HTTPException, record_trace) are gone
with them. resolve_livedocs.py now imports directly from live_docs_fetch.
